requirements-dev.txt
django-seed>=0.3